:label:Rename: 동행글 기능별 파일 분리 #37
1. acommpany.js에는 동행글 조회 및 삭제 함수가 있습니다.
2. accompany-posting.js에는 동행글 작성 및 관련 함수가 있습니다.
3. accompany-editing.js에는 동행글 수정 및 관련 함수가 있습니다.
